backend/pyproject.toml
[tool.poetry]
name = "ceezaa-backend"
version = "0.1.0"
description = "Ceezaa MVP Backend - Taste Intelligence Layer"
authors = ["Ceezaa Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
supabase = "^2.3.0"
plaid-python = "^16.0.0"
openai = "^1.10.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.0"
black = "^24.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "exploration: marks tests as exploration tests (for understanding external APIs)",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
